refactor(color): change `Rgb` struct type to a float instead of a byte

+++ pkg/color/color.go
@@ -1,5 +1,5 @@
 package color
 
 type Rgb struct {
-	R, G, B byte
+	R, G, B float32
 }
